add encryptinc
