+++ backend/encrypting/encrypting_mk.py
@@ -1,0 +1,20 @@
+import base64
+import os
+from cryptography.fernet import Fernet
+from cryptography.hazmat.primitives import hashes
+from cryptography.hazmat.primitives.kdf.pbkdf2 import PBKDF2HMAC
+
+
+def encrypting_mk(master_key, salt=''):
+    if salt == '':
+        salt = os.urandom(16)
+    kdf = PBKDF2HMAC(
+        algorithm=hashes.SHA256(),
+        length=32,
+        salt=salt,
+        iterations=100_000,
+    )
+    key = base64.urlsafe_b64encode(kdf.derive(master_key.encode()))
+    f = Fernet(key)
+    encrypted_control_string = f.encrypt(b"control")
+    return encrypted_control_string, salt
