+++ pyproject.toml
@@ -5,6 +5,7 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
+    "cryptography>=44.0.2",
     "psycopg2>=2.9.10",
     "pyqt6>=6.4.2",
     "pyqt6-tools>=6.4.2.3.3",
